refactor: move relabelling stuff to its own file
refactor: move relabelling stuff to its own file

refactor: move test too
