+++ CreeDictionary/API/search/presentation.py
@@ -6,7 +6,8 @@ from django.forms import model_to_dict
 
 from utils import get_modified_distance
 from . import types, core, lookup
-from utils.fst_analysis_parser import LABELS, partition_analysis
+from utils.fst_analysis_parser import partition_analysis
+from CreeDictionary.relabelling import LABELS
 from utils.types import FSTTag, Label, ConcatAnalysis
 from .types import Preverb, LinguisticTag, linguistic_tag_from_fst_tags
 from ..models import Wordform
