+++ CreeDictionary/tests/CreeDictionary_tests/test_relabelling.py
@@ -4,7 +4,7 @@
 from io import StringIO
 
 import pytest
-from utils.fst_analysis_parser import Relabelling
+from CreeDictionary.relabelling import Relabelling
 
 labels = Relabelling.from_tsv(
     StringIO(
